# CITFAC-openapi.yaml
openapi: 3.0.3
info:
  title: Città Facile
  description: Insieme di API esposte dal BFF di Città Facile (nativamente o attraverso proxy di API esistenti)
  contact:
    email: maurizio.peisino@csi.it
  version: 1.0.0
servers:
  - url: 'https://localhost'
...
paths:
...
  '/api/preferenze/{service-id}':
    parameters:
      - $ref: '#/components/parameters/idServizio'
    post:
      operationId: addPreferenza
      summary: Add new preference
      description: Add new  preferences of a user for a given tenant
      tags:
        - Authenticated | Preferences
      responses:
        '200':
          description: Preferenza aggiunta con Successo
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  example: 2
        '400':
          $ref: '#/components/responses/400'
    delete:
      operationId: removePreferenza
      summary: Delete preference
      description: Delete a preference of a user for a given tenant
      tags:
        - Authenticated | Preferences
      responses:
        '200':
          description: Preferenza eliminata con Successo
          content:
            application/json:
              example:
                status: 200
                title: Preferenza eliminata con Successo
        '400':
          $ref: '#/components/responses/400'
...
  /api-public/utenti/me:
    get:
      operationId: getCurrentUser
      summary: Get current user
      description: Get user details based on data provided by shibboleth
      tags:
        - Public | Current user
      responses:
        '200':
          description: Current user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utente'
        '401':
          $ref: '#/components/responses/401'
...
  '/api-admin/tenant/{tenant-id}':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getTenant
      summary: Get tenant
      description: Get single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
    put:
      operationId: putTenant
      summary: Update tenant
      description: Update single tenant configuration
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
    patch:
      operationId: patchTenant
      summary: Update tenant partially
      description: Update single tenant configuration passing modified fields only
      tags:
        - Administration | Tenants
      responses:
        '200':
          description: Updated tenant configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComuneAdmin'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '422':
          $ref: '#/components/responses/422'
  '/api-admin/tenant/{tenant-id}/services':
    parameters:
      - $ref: '#/components/parameters/idComune'
    get:
      operationId: getServices
      summary: Get all services
      description: Get all services for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Servizio'
    delete:
      operationId: deleteServices
      summary: Delete all services
      description: Delete all services for single tenant
      tags:
        - Administration | Services
      responses:
        '200':
          description: Services list is deleted
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
...
components:
  parameters:
    idComune:
      name: tenant-id
      in: path
      required: true
      schema:
        type: integer
        default: 0
    idServizio:
      name: service-id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    '400':
      description: Bad Request. Body or parameters validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '401':
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '403':
      description: Forbidden. User have no access to resource or tenant is not corresponding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '404':
      description: Not Found. Cannot find entity key or tenant id related to the entity
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    '422':
      description: Unprocessable Entity. Duplication of unique fields or other integrity constrain violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
...
  schemas:
    Errore:
      type: object
      properties:
        status:
          type: integer
          format: int32
        title:
          type: string
          example: Messaggio di errore
        detail:
          type: array
          nullable: true
          items:
            type: string
            example: 'Nome campo: Dettaglio messaggio di errore'
    Utente:
      type: object
      properties:
        nome:
          type: string
        cognome:
          type: string
        codFisc:
          type: string
          format: codice-fiscale
        contatti:
          type: object
          properties:
            sms:
              type: string
              format: phone
              example: 00393807129011
            email:
              type: string
              format: email
            push:
              type: object
...
    Comune:
      type: object
      properties:
        nome:
          type: string
          example: Novara
        regione:
          $ref: '#/components/schemas/Regione'
        tenant:
          type: string
          example: comune.pinerolo.to.it
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        homePage:
          type: string
          example: 'https://www.comune.pinerolo.to.it'
        codiceEnte:
          type: string
          description: codice IPA del comune
          example: c_g388
    ComuneAdmin:
      allOf:
        - $ref: '#/components/schemas/Comune'
        - type: object
          properties:
            id:
              type: integer
              example: 1
            templateNotificatore:
              type: string
            logoutUri:
              type: string
            cf:
              type: string
            cms:
              type: object
              properties:
                urlServizi:
                  type: string
                urlMenu:
                  type: string
                backOffice:
                  type: string
            assets:
              type: array
              items:
                type: object
                properties:
                  tipoCredenziali:
                    type: string
                  urlBackoffice:
                    type: string
                  icon:
                    type: string
            apiManagerKey:
              type: string
            apiManagerSecret:
              type: string
    Regione:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Piemonte
        link:
          type: string
          format: uri
          description: link al portale regionale
          example: 'https://www.regione.piemonte.it/'
    Servizio:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Nome servizio
        descrizione:
          type: string
          example: Descrizione servizio
        categoriaServizio:
          type: object
          properties:
            id:
              type: integer
              example: 1
            nome:
              type: string
              example: Categoria servizio
        linkScheda:
          type: string
          format: uri
          description: link alla scheda descrittiva del servizio
        linkServizio:
          type: string
          format: uri
          description: link del servizio
        servizioCittaFacile:
          type: boolean
          description: indica se il servizio viene lanciato da città facile
          example: true
        servizioAttivo:
          type: boolean
          description: indica se il servizio è attivo oppure in manutenzione
          example: true
        messaggioManutenzione:
          type: string
          nullable: true
          description: messaggio da mostrare all'utente quando il servizio risulta non attivo
          example: Il servizio non è disponibile per attività di manutenzione straordinaria
        notificabile:
          type: boolean
          example: false
        codiceNotifiche:
          type: string
          nullable: true
        assistenza:
          type: boolean
          example: false
        codice_assistenza:
          type: string
          nullable: true
...
    Asset:
      type: object
      properties:
        id:
          type: integer
          example: 1
        codice:
          type: string
          example: PPAY
        nome:
          type: string
          example: PiemontePay
        descrizione:
          type: string
          example: Gestione dei pagamenti per comuni del piemonte
        urlBackoffice:
          type: string
          nullable: true
          example: 'null'
        icon:
          type: string
          nullable: true
          example: 'null'
...